Convert record counts to long
In large datsets, 32 bit integers can be insufficient to counts the
number of reads.
This commit replaces it by longs.

Reported by @jamesnemesh.

+++ src/java/org/broadinstitute/dropseqrna/utils/PassFailTrackingIteratorI.java
@@ -7,7 +7,7 @@ package org.broadinstitute.dropseqrna.utils;
  */
 public interface PassFailTrackingIteratorI {
 
-	public int getRecordsPassed();
-	public int getRecordsFailed();
+	public long getRecordsPassed();
+	public long getRecordsFailed();
 	
 }
